rename category to service
