+++ application/__init__.py
@@ -15,7 +15,7 @@ def create_app():
         user_util.add_new_user('username', 'password', True)
         rate_util.add_new({'name': 'Basic', 'speed': 50, 'price': 100, 'description': "Basic rate"})
         rate_util.add_new({'name': 'Basic1', 'speed': 41, 'price': 24, 'description': "Standart rate"})
-        category_util.add_new({'name': 'category'})
+        category_util.add_new({'name': 'service'})
         news_util.add_news({'title': 'Test new', 'description': 'Test news'})
         news_util.add_news({'title': 'Test new1', 'description': 'Test news'})
 
@@ -31,8 +31,8 @@ def create_app():
     from application.routes.news_routes import bp as news_bp
     app.register_blueprint(news_bp, url_prefix='/news')
 
-    from application.routes.category_routes import bp as category_bp
-    app.register_blueprint(category_bp, url_prefix='/category')
+    from application.routes.service_routes import bp as service_bp
+    app.register_blueprint(service_bp, url_prefix='/service')
 
     return app
