+++ application/routes/service_routes.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, render_template, redirect, request
 from application.util import rate_util
 
-bp = Blueprint('category', __name__)
+bp = Blueprint('service', __name__)
 
 
 @bp.route('/')
-def show_categories():
-    return render_template('category/category.html')
+def show_services():
+    return render_template('service/service.html')
