+++ application/templates/base.html
@@ -10,7 +10,7 @@
 
     <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/news.css') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/category.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/service.css') }}">
     <link rel="stylesheet" href="{{ url_for('static', filename='css/rate.css') }}">
 
     <link rel="apple-touch-icon" sizes="180x180" href="{{ url_for('static', filename='img/apple-touch-icon.png') }}">
@@ -41,7 +41,7 @@
                 <a class="nav-link" href="{{ url_for('news.show_news') }}">News</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('category.show_categories') }}">Category</a>
+                <a class="nav-link" href="{{ url_for('service.show_services') }}">Service</a>
             </li>
         </ul>
     </div>
@@ -82,7 +82,7 @@
                 <li class="footer__list-item"><a href="#" class="footer__link">About Us</a></li>
                 <li class="footer__list-item"><a href="{{ url_for('rate.show_rates') }}" class="footer__link">Rate</a>
                 </li>
-                <li class="footer__list-item"><a href="{{ url_for('category.show_categories') }}" class="footer__link">Category</a>
+                <li class="footer__list-item"><a href="{{ url_for('service.show_services') }}" class="footer__link">Service</a>
                 </li>
                 <li class="footer__list-item"><a href="{{ url_for('news.show_news') }}" class="footer__link">News</a>
                 </li>
